requirements.txt
certifi==2017.4.17
chardet==3.0.3
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.2
django-braces==1.11.0
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.6
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.1
openapi-codec==1.3.1
pycryptodome==3.4.6
PyJWT==1.5.0
python3-openid==3.1.0
pytz==2017.2
requests==2.17.3
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
uritemplate==3.0.0
urllib3==1.21.1
